+++ webapp/src/components/rhs/rhs_scrollbars.tsx
@@ -1,0 +1,44 @@
+import React from 'react';
+
+import Scrollbars from 'react-custom-scrollbars';
+
+function renderView(props: any) {
+    return (
+        <div
+            {...props}
+            className='scrollbar--view'
+        />);
+}
+
+function renderThumbHorizontal(props: any) {
+    return (
+        <div
+            {...props}
+            className='scrollbar--horizontal'
+        />);
+}
+
+function renderThumbVertical(props: any) {
+    return (
+        <div
+            {...props}
+            className='scrollbar--vertical'
+        />);
+}
+
+const RHSScrollbars = ({children}: {children: React.ReactNode[]}) => {
+    return (
+        <Scrollbars
+            autoHide={true}
+            autoHideTimeout={500}
+            autoHideDuration={500}
+            renderThumbHorizontal={renderThumbHorizontal}
+            renderThumbVertical={renderThumbVertical}
+            renderView={renderView}
+        >
+            {children}
+        </Scrollbars>
+    );
+};
+
+export default RHSScrollbars;
